+++ examples/inverse/plot_morph_volume_stc.py
@@ -6,14 +6,23 @@ Morph volumetric source estimate
 This example demonstrates how to morph an individual subject's
 :class:`mne.VolSourceEstimate` to a common reference space. We achieve this
 using :class:`mne.SourceMorph`. Pre-computed data will be morphed based on
-an affine transformation and a nonlinear morph, estimated on the respective
-transformation from the subject's anatomical T1 weighted MRI (brain) to
-FreeSurfer's 'fsaverage' T1 weighted MRI (brain).See
+an affine transformation and a nonlinear registration method
+known as Symmetric Diffeomorphic Registration (SDR) by Avants et al. [1]_.
+
+Transformation is estimated from the subject's anatomical T1 weighted MRI
+(brain) to FreeSurfer's 'fsaverage' T1 weighted MRI (brain). See
 https://surfer.nmr.mgh.harvard.edu/fswiki/FsAverage .
 
 Afterwards the transformation will be applied to the volumetric source
 estimate. The result will be plotted, showing the fsaverage T1 weighted
 anatomical MRI, overlaid with the morphed volumetric source estimate.
+
+References
+----------
+.. [1] Avants, B. B., Epstein, C. L., Grossman, M., & Gee, J. C. (2009).
+       Symmetric Diffeomorphic Image Registration with Cross- Correlation:
+       Evaluating Automated Labeling of Elderly and Neurodegenerative
+       Brain, 12(1), 26-41.
 
 .. note:: For a tutorial about morphing see:
           :ref:`sphx_glr_auto_tutorials_plot_morph_stc.py`.